Create an EmbeddedPaymentElementController
