+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Embedded/EmbeddedPaymentElement.swift
@@ -19,7 +19,11 @@ public class EmbeddedPaymentElement {
     public let view: UIView
 
     /// A view controller to present on.
-    public var presentingViewController: UIViewController?
+    public var presentingViewController: UIViewController? {
+        didSet {
+            embeddedController.presentingViewController = presentingViewController
+        }
+    }
 
     /// This contains the `configuration` you passed in to `create`.
     public let configuration: Configuration
@@ -45,12 +49,10 @@ public class EmbeddedPaymentElement {
 
     /// The customer's currently selected payment option.
     public var paymentOption: PaymentOptionDisplayData? {
-        return embeddedPaymentMethodsView.displayData
+        return embeddedController.embeddedPaymentMethodsView.displayData
     }
-
-    private let embeddedPaymentMethodsView: EmbeddedPaymentMethodsView
-    private let loadResult: PaymentSheetLoader.LoadResult
-    private let analyticsHelper: PaymentSheetAnalyticsHelper
+    
+    private let embeddedController: EmbeddedPaymentElementController
 
     /// An asynchronous failable initializer
     /// This loads the Customer's payment methods, their default payment method, etc.
@@ -62,65 +64,8 @@ public class EmbeddedPaymentElement {
         intentConfiguration: IntentConfiguration,
         configuration: Configuration
     ) async throws -> EmbeddedPaymentElement {
-        // TODO(porter) Should we create a new analytics helper specific to embedded? Figured this out when we do analytics.
-        let analyticsHelper = PaymentSheetAnalyticsHelper(isCustom: true, configuration: PaymentSheet.Configuration())
-        AnalyticsHelper.shared.generateSessionID()
-
-        let loadResult = try await PaymentSheetLoader.load(mode: .deferredIntent(intentConfiguration),
-                                                           configuration: configuration,
-                                                           analyticsHelper: analyticsHelper,
-                                                           integrationShape: .embedded)
-
-        let paymentMethodTypes = PaymentSheet.PaymentMethodType.filteredPaymentMethodTypes(from: .deferredIntent(intentConfig: intentConfiguration),
-                                                                                           elementsSession: loadResult.elementsSession,
-                                                                                           configuration: configuration,
-                                                                                           logAvailability: true)
-        let shouldShowApplePay = PaymentSheet.isApplePayEnabled(elementsSession: loadResult.elementsSession, configuration: configuration)
-        let shouldShowLink = PaymentSheet.isLinkEnabled(elementsSession: loadResult.elementsSession, configuration: configuration)
-        let savedPaymentMethodAccessoryType = await RowButton.RightAccessoryButton.getAccessoryButtonType(
-            savedPaymentMethodsCount: loadResult.savedPaymentMethods.count,
-            isFirstCardCoBranded: loadResult.savedPaymentMethods.first?.isCoBrandedCard ?? false,
-            isCBCEligible: loadResult.elementsSession.isCardBrandChoiceEligible,
-            allowsRemovalOfLastSavedPaymentMethod: configuration.allowsRemovalOfLastSavedPaymentMethod,
-            allowsPaymentMethodRemoval: loadResult.elementsSession.allowsRemovalOfPaymentMethodsForPaymentSheet()
-        )
-
-        let initialSelection: EmbeddedPaymentMethodsView.Selection? = {
-            // Default to the customer's default or the first saved payment method, if any
-            let customerDefault = CustomerPaymentOption.defaultPaymentMethod(for: configuration.customer?.id)
-            switch customerDefault {
-            case .applePay:
-                return .applePay
-            case .link:
-                return .link
-            case .stripeId, nil:
-                return loadResult.savedPaymentMethods.first.map { .saved(paymentMethod: $0) }
-            }
-        }()
-
-        let embeddedPaymentMethodsView = await EmbeddedPaymentMethodsView(
-            initialSelection: initialSelection,
-            paymentMethodTypes: paymentMethodTypes,
-            savedPaymentMethod: loadResult.savedPaymentMethods.first,
-            appearance: configuration.appearance,
-            shouldShowApplePay: shouldShowApplePay,
-            shouldShowLink: shouldShowLink,
-            savedPaymentMethodAccessoryType: savedPaymentMethodAccessoryType,
-            mandateProvider: VerticalListMandateProvider(configuration: configuration,
-                                                         elementsSession: loadResult.elementsSession,
-                                                         intent: .deferredIntent(intentConfig: intentConfiguration)),
-            shouldShowMandate: configuration.embeddedViewDisplaysMandateText
-        )
-
-        let embeddedPaymentElement: EmbeddedPaymentElement = .init(view: embeddedPaymentMethodsView,
-                                                                   configuration: configuration,
-                                                                   loadResult: loadResult,
-                                                                   analyticsHelper: analyticsHelper)
-        await MainActor.run {
-            embeddedPaymentMethodsView.delegate = embeddedPaymentElement
-        }
-
-        return embeddedPaymentElement
+        let embeddedPaymentElementController: EmbeddedPaymentElementController = try await .create(intentConfiguration: intentConfiguration, configuration: configuration)
+        return .init(view: embeddedPaymentElementController.embeddedPaymentMethodsView, configuration: configuration, embeddedController: embeddedPaymentElementController)
     }
 
     /// The result of an `update` call
@@ -155,13 +100,12 @@ public class EmbeddedPaymentElement {
 
     // MARK: - Internal
 
-    private init(view: EmbeddedPaymentMethodsView, configuration: Configuration, loadResult: PaymentSheetLoader.LoadResult, analyticsHelper: PaymentSheetAnalyticsHelper, delegate: EmbeddedPaymentElementDelegate? = nil) {
+    private init(view: EmbeddedPaymentMethodsView, configuration: Configuration, embeddedController: EmbeddedPaymentElementController, delegate: EmbeddedPaymentElementDelegate? = nil) {
         self.view = view
-        self.embeddedPaymentMethodsView = view
         self.delegate = delegate
         self.configuration = configuration
-        self.loadResult = loadResult
-        self.analyticsHelper = analyticsHelper
+        self.embeddedController = embeddedController
+        self.embeddedController.delegate = self
     }
 }
 
@@ -242,71 +186,16 @@ extension EmbeddedPaymentElement {
     public typealias ExternalPaymentMethodConfiguration = PaymentSheet.ExternalPaymentMethodConfiguration
 }
 
-extension EmbeddedPaymentElement: EmbeddedPaymentMethodsViewDelegate {
+extension EmbeddedPaymentElement: EmbeddedPaymentElementControllerDelegate {
     func heightDidChange() {
         delegate?.embeddedPaymentElementDidUpdateHeight(embeddedPaymentElement: self)
     }
-
-    func selectionDidUpdate() {
+    
+    func selectionDidChange() {
         delegate?.embeddedPaymentElementDidUpdatePaymentOption(embeddedPaymentElement: self)
-        guard case let .new(paymentMethodType) = embeddedPaymentMethodsView.selection else {
-            // If the selection is not `.new`, e.g. `link`, `applePay`, or `saved` we do not need to show a form.
-            return
-        }
-        
-        guard let presentingViewController = presentingViewController else {
-            assertionFailure("Presenting view controller not found, set EmbeddedPaymentElement.presentingViewController.")
-            return
-        }
-        
-        let embeddedFormVC = EmbeddedFormViewController(
-            configuration: configuration,
-            loadResult: loadResult,
-            paymentMethodType: paymentMethodType,
-            analyticsHelper: analyticsHelper
-        )
-        embeddedFormVC.delegate = self
-        
-        // Only show forms that require user input
-        guard embeddedFormVC.collectsUserInput else { return }
-        
-        let bottomSheet = BottomSheetViewController(
-            contentViewController: embeddedFormVC,
-            appearance: configuration.appearance,
-            isTestMode: configuration.apiClient.isTestmode,
-            didCancelNative3DS2: {} // TODO(porter) Cancel 3DS2 on the payment handler
-        )
-        
+    }
+    
+    func willPresentForm() {
         delegate?.embeddedPaymentElementWillPresent(embeddedPaymentElement: self)
-        presentingViewController.presentAsBottomSheet(bottomSheet, appearance: configuration.appearance)
-    }
-
-}
-
-extension EmbeddedPaymentElement: EmbeddedFormViewControllerDelegate {
-    func embeddedFormViewControllerShouldConfirm(_ embeddedFormViewController: EmbeddedFormViewController,
-                                                 with paymentOption: PaymentOption,
-                                                 completion: @escaping (PaymentSheetResult, STPAnalyticsClient.DeferredIntentConfirmationType?) -> Void) {
-        // TODO(porter) Finish confirmation
-    }
-    
-    func embeddedFormViewControllerDidFinish(_ embeddedFormViewController: EmbeddedFormViewController, result: PaymentSheetResult) {
-        embeddedFormViewController.dismiss(animated: true) { [weak self] in
-            guard case let .confirm(completion) = self?.configuration.formSheetAction else {
-                return
-            }
-            
-            completion(result)
-        }
-    }
-    
-    func embeddedFormViewControllerDidCancel(_ embeddedFormViewController: EmbeddedFormViewController) {
-        embeddedFormViewController.dismiss(animated: true)
-        // TODO(porter) Notify formSheet completion handler?
-    }
-    
-    func embeddedFormViewControllerShouldClose(_ embeddedFormViewController: EmbeddedFormViewController) {
-        // TOOD(porter) Handle dismiss
-        embeddedFormViewController.dismiss(animated: true)
     }
 }
