+++ StripePaymentSheet/StripePaymentSheet/Source/PaymentSheet/Embedded/EmbeddedPaymentMethodsView.swift
@@ -12,7 +12,7 @@ import UIKit
 
 protocol EmbeddedPaymentMethodsViewDelegate: AnyObject {
     func heightDidChange()
-    func selectionDidUpdate()
+    func selectionTapped(didChange: Bool)
 }
 
 /// The view for an embedded payment element
@@ -29,9 +29,7 @@ class EmbeddedPaymentMethodsView: UIView {
     private(set) var selection: Selection? {
         didSet {
             updateMandate()
-            if oldValue != selection {
-                delegate?.selectionDidUpdate()
-            }
+            delegate?.selectionTapped(didChange: oldValue != selection)
         }
     }
     private let mandateProvider: MandateTextProvider
